requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
motor==3.3.1
pymongo==4.5.0
pydantic==2.4.2
python-jose==3.3.0
python-dotenv==1.0.0
email-validator==2.0.0
python-multipart==0.0.6
cryptography==40.0.2
google-auth==2.22.0
requests==2.31.0
passlib>=1.7.4,<2.0
bcrypt==3.2.0
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.11.1
httpx==0.25.0
mongomock==4.1.2
pytest-env==1.0.1